Functions Variables

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	double a, b, c;
+	float a, b, c;
 	int count;
 
 	a = 1;
